refactor: isolate old fiat on ramp
