+++ .eslintrc.js
@@ -76,10 +76,10 @@ const allowedBarrelFiles = [
   'src/resources/f2c/index.ts',
   'src/resources/nfts/index.ts',
   'src/resources/nfts/simplehash/index.ts',
-  'src/screens/AddCash/index.tsx',
   'src/screens/Diagnostics/index.tsx',
   'src/screens/expandedAssetSheet/components/sections/index.tsx',
   'src/screens/Explain/index.tsx',
+  'src/screens/FiatOnRamp/index.tsx',
   'src/screens/NFTOffersSheet/index.tsx',
   'src/screens/NFTSingleOfferSheet/index.tsx',
   'src/screens/NotificationsPromoSheet/index.tsx',
