+++ src/features/cash/navigation/getAddCashRoute.ts
@@ -1,7 +1,7 @@
 import Routes from '@/navigation/routesNames';
 
-type AddCashRoute = typeof Routes.CASH_DEPOSIT_INTRO_PANEL | typeof Routes.ADD_CASH_SHEET;
+type AddCashRoute = typeof Routes.CASH_DEPOSIT_INTRO_PANEL | typeof Routes.FIAT_ON_RAMP_SHEET;
 
 export function getAddCashRoute(isCashEnabled: boolean): AddCashRoute {
-  return isCashEnabled ? Routes.CASH_DEPOSIT_INTRO_PANEL : Routes.ADD_CASH_SHEET;
+  return isCashEnabled ? Routes.CASH_DEPOSIT_INTRO_PANEL : Routes.FIAT_ON_RAMP_SHEET;
 }
