+++ src/screens/FiatOnRamp/utils.ts
@@ -1,4 +1,4 @@
-import { Network as APINetwork } from '@/screens/AddCash/types';
+import { Network as APINetwork } from '@/screens/FiatOnRamp/types';
 import { ChainId } from '@/state/backendNetworks/types';
 
 export function convertAPINetworkToInternalChainIds(network: APINetwork): ChainId | undefined {
